Remove some transaction log validation for schema changes
The checks for things which are valid for additive-only but not automatic
Realms was not being consistently applied, and fixing this isn't really worth
the effort as everything either works or fails in a sensible way anyway. The
checks for destructive changes remain as they actually provide some value
(without the validation, trying to use a column which was deleted by a
different process would produce mysterious crashes not obviously related to the
root problem), and are already being always used.