Add proper support for opening a Realm without a schema and then with a schema
The GN needs to open Realms without specifying a schema, while in the same
process the user needs to be able to open the same Realm, on the same
thread/execution context and specify a schema. The code previously assumed that
this switch would never happen, and so it would either throw an exception or
outright ignore the supplied schema, depending on the sequence things were
supplied and whether or not caching was enabled.