Fix an issue with dying sessions that are revived
When a session is created, a bunch of lambdas are created that capture a weak pointer from the `shared_ptr` to the `SyncSession`. Those weak pointers become invalid as soon as the conditions that make the session enter `dying` are met. If the session is revived and returned to service, those weak pointers remain invalid forever. This changes the lambdas to try and update their weak pointers if the weak pointers are invalid.