Adjust how property types are defined to support arrays of all types
This decouples the values of PropertyType from the corresponding core
enums, changes PropertyType::Array into a flag which can be combined
with any of the other property types and adds a PropertyType::Nullable
flag to replace the separate bool.

This makes it possible to express property types not natively supported
by core (such as the upcoming arrays of primitives), and makes it so
that the property type is fully expressed in a single value.

The definitions of properties within the tests and sync code have also
been changed to use tagged bools rather than an unreadable and error
prone sequence of `true, true, false` for setting the primary key and
indexed statuses.