Allow changing primary keys in migrations
Primary keys need to be modifiable within the migration function to support
adding a new primary key column to an existing table or changing the type of
the primary key. This does not violate sync's restriction that primary keys
cannot be modified because migration functions are not supported for
synchronized realms.