Build coverage-enabled tests with CATCH_CONFIG_FAST_COMPILE
This improves compile time (159s -> 145s on my machine) and more importantly
makes the object files less unreasonably big (1.2 GB total -> 850 MB) which
should also reduce the amount of ram needed to link the executable, which has
been a problem in the past.

This comes at the cost of making error reporting when unexpected exceptions are
thrown significantly worse, but that doesn't really matter for coverage runs.