Make ExternalCommitHelper robust on Windows
Starting with Core 3.0 we have a robust interprocess mutex and condition variable implementation on Windows we can use to reliably notify interested parties of commits at the Object Store level.

The implementation heavily borrows from `SharedGroup::wait_for_change()`. In fact, I believe the generic ExternalCommitHelper implementation should perform equally well on Windows - we’d be using it instead of this one if it only worked with synced realms.