Rename update_tableveiew and expose it to public
Normally size() is much faster than creating the tableview. But in some
cases, like string relevant queries, the size() and creating the
tableview may take similar time to exceed.

The orignal purpose of not creating TableView in size() is to make the
call faster if the query has not been executed before. But in the real
usecases, users always try to call accessors after the size() which
makes the query gets executed twice. See java issue
https://github.com/realm/realm-java/issues/5328
https://github.com/realm/realm-java/issues/5387

By exposing evaluate_query_if_needed(), binding could give user a chance
to execute the query immediately to avoid this.