Create the sync session in set_transaction_callback()
If it doesn't already exist.

The global notifier opens and then immediately closes each Realm during startup
purely for the sake of creating the sync session (it used to do more with the
Realms, but hasn't for a long time), and with a large number of Realms this
starts to take a significant amount of time.