Differnt change set behavior for list and results
If there is no actual changes in the list like remove all elements from
list and added them again in the same transaction, the Results built
from List can detect there is no change actually happened thus no
notifications will be triggered. But in the same case, notification will
be triggered for the List with a change set of deletions and insertions.

Not sure if this is by design.