Remove bind_with_admin_token() and use the normal refresh flow
When session multiplexing is enabled we need to perform a token refresh even
with admin token users so that ROS knows that we're accessing a new Realm (as
it can't just automatically detect that from the fact that we've opened a new
connection).

All bindings need to be updated to support this, but bindings which do not
support multiplexing (i.e. everything but JS) can simply "refresh" the token on
the session without talking to the server.