Fix a crash when an observed primitive list is deleted before the first run of the notifier
For version management reasons, the first run of a notifier is done on a
different SharedGroup from all of the subsequent runs. This requires detaching
the notifier from the first SG and attaching it to the second one by handing
over the Table. This doesn't work if the Table was deleted, so we need to
explicitly handle that case.