Fix "previous <= m_schema_transaction_version_max" assertion failure
Example code which hit this:

    std::vector<ObjectSchema> classes;
    for (int i = 0; i < 10'000; ++i) {
        classes.push_back({util::format("type %1", i), {
            {"value", PropertyType::Int}
        }});
    }
    InMemoryTestFile config;
    config.cache = false;
    config.automatic_change_notifications = false;
    auto r = Realm::get_shared_realm(config);
    r->update_schema(classes);

    std::thread t([&] { Realm::get_shared_realm(config); });
    while (true) {
        r->begin_transaction(); r->commit_transaction();
    }

Previously Realm::update_schema() would clear RealmCoordinator's schema cache,
but not repopulate it. This was because populating it requires rereading it
from the Realm, so it was optimal to defer doing that until it was next needed
(i.e. the next time a Realm instance was obtained). However, this lead to a
race condition: if the Realm which performed the migration committed another
write while a different Realm was in the middle of reading the schema
(specifically in between acquiring the read lock and caching the read schema),
the schema cache would be put into an invalid state, as the new Realm would
cache the schema just for that older version.

The simplest fix for this is to just reread the schema after preforming a
migration so that caching works normally. This has a small performance hit, but
it's very unlikely to ever matter compared to the cost of actually performing
schema changes.