Fix sync metadata key lookup failures
With previous versions we stored the key with `kSecAttrAccessibleAlways`, and
looking it up with `kSecAttrAccessibleAfterFirstUnlock` sometimes doesn't work.
Fortunately there's no reason to actually include it in the search dictionary
for lookup, and only for setting.

I was unable to write a test which reproduces the problem; storing with Always
and then immediately looking up with AfterFirstUnlock works unfortunately.