Don't pass bogus length to string_view::substr()
This would never actually change behavior since it would always pass a length
larger than desired which would automatically be interpreted as stopping at
the end of the string. But it seems better to either pass the real desired
length, or use the default of npos. This was an issue in the JS port of this
code for realm-web, and I'd prefer to keep the differences between the
implementations to a minimum.